+++ .github/workflows/deploy_documents.yml
@@ -1,0 +1,32 @@
+name: Deploy documents
+on:
+  push:
+    branches: [ main ]
+    paths:
+      - 'lib/**'
+      - '**.md'
+      - '**.txt'
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: ruby/setup-ruby@v1
+        with:
+          ruby-version: '3.0'
+          bundler-cache: true # runs 'bundle install' and caches installed gems automatically
+      - name: git setting
+        run: |
+          git config --local user.email "kachick1@gmail.com"
+          git config --local user.name "Kenichi Kamiya"
+      - name: Update YARD documents and deploy to GitHub Pages
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          DATE=`date +"%Y%m%d-%H%M-UTC"`
+          BRANCH_NAME="documents-$DATE"
+          git checkout -b "$BRANCH_NAME"
+          bundle exec yard --output-dir docs
+          git add docs
+          git commit -m 'Update YARD documents' docs
+          git push origin "$BRANCH_NAME":gh-pages --force
