+++ .github/workflows/test_behaviors.yml
@@ -1,0 +1,41 @@
+# https://docs.github.com/ja/actions/reference/workflow-syntax-for-github-actions
+name: Test
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - 'lib/**'
+      - 'test/**'
+      - 'spec/**'
+      - '.rspec'
+      - '**.gemspec'
+      - 'Gemfile'
+      - 'Rakefile'
+  pull_request:
+    paths:
+      - 'lib/**'
+      - 'test/**'
+      - 'spec/**'
+      - '.rspec'
+      - '**.gemspec'
+      - 'Gemfile'
+      - 'Rakefile'
+jobs:
+  test_behaviors:
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest]
+        # Due to https://github.com/actions/runner/issues/849, we have to use quotes for '3.0'
+        ruby: [2.6, 2.7, '3.0', head]
+    runs-on: ${{ matrix.os }}
+    steps:
+    - uses: actions/checkout@v2
+    - uses: ruby/setup-ruby@v1
+      env:
+        BUNDLE_WITHOUT: development
+      with:
+        ruby-version: ${{ matrix.ruby }}
+        bundler-cache: true # runs 'bundle install' and caches installed gems automatically
+    - run: bundle exec rake test_behaviors --suppress-backtrace='\/bundle'
