+++ .github/workflows/test_signatures.yml
@@ -1,0 +1,44 @@
+name: Signature
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - 'lib/**'
+      - 'sig/**'
+      - 'exe/**'
+      - '**.md'
+      - '**.txt'
+      - '**.gemspec'
+      - 'Gemfile'
+      - 'Steepfile'
+      - '.yardopts'
+      - 'Rakefile'
+  pull_request:
+    paths:
+      - 'lib/**'
+      - 'sig/**'
+      - 'exe/**'
+      - '**.md'
+      - '**.txt'
+      - '**.gemspec'
+      - 'Gemfile'
+      - 'Steepfile'
+      - '.yardopts'
+      - 'Rakefile'
+jobs:
+  validate_signatures:
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest]
+        # Due to https://github.com/actions/runner/issues/849, we have to use quotes for '3.0'
+        ruby: ['3.0']
+    runs-on: ${{ matrix.os }}
+    steps:
+    - uses: actions/checkout@v2
+    - uses: ruby/setup-ruby@v1
+      with:
+        ruby-version: ${{ matrix.ruby }}
+        bundler-cache: true # runs 'bundle install' and caches installed gems automatically
+    - run: bundle exec rake validate_signatures >/dev/null
