--- .travis.yml
@@ -1,17 +1,0 @@
-language: ruby
-
-rvm:
-  - ruby-head
-  - 2.6
-  - 2.5
-  - rbx-3
-  - jruby-head
-
-matrix:
-  allow_failures:
-    - rvm: ruby-head
-    - rvm: rbx-3
-    - rvm: jruby-head
-
-before_install:
-  - gem install bundler
