+++ .yardopts
@@ -1,0 +1,3 @@
+--fail-on-warning
+-
+MIT-LICENSE.txt
