+++ Gemfile
@@ -1,12 +1,26 @@
+# coding: us-ascii
+# frozen_string_literal: true
+
 source 'https://rubygems.org'
 
 gemspec
 
+group :development, :test do
+  gem 'rake', '>= 13.0.3', '< 20.0'
+  gem 'irb', '>= 1.3.5', '< 2.0'
+  gem 'irb-power_assert', '0.0.2'
+end
+
 group :development do
-  gem 'rake'
+  gem 'yard', '>= 0.9.26', '< 2', require: false
+  gem 'rubocop', '>= 1.16.0', '< 1.17.0', require: false
+  gem 'rubocop-rake', '>= 0.5.1', '< 0.6.0', require: false
+  gem 'rubocop-performance', '>= 1.11.3', '< 1.12.0', require: false
+  gem 'rubocop-rubycw', '>= 0.1.6', '< 0.2.0', require: false
+  gem 'rubocop-md', '>= 1.0.1', '< 2.0.0', require: false
 end
 
 group :test do
-  gem 'rake'
-  gem 'declare', '~> 0.0.5.c'
+  gem 'declare', '>= 0.3.1', '< 0.4.0'
+  gem 'warning', '>= 1.2.0', '< 2.0'
 end
