+++ net-ipaddress.gemspec
@@ -1,33 +1,45 @@
 # coding: us-ascii
+# frozen_string_literal: true
 
-require File.expand_path('../lib/net/ipaddress/version', __FILE__)
+lib_name = 'net-ipaddress'
+
+require_relative './lib/net/ipaddress/version'
+repository_url = "https://github.com/kachick/#{lib_name}"
 
 Gem::Specification.new do |gem|
-  # specific
-
-  gem.description   = %q{Utils for ipaddresses}
-  gem.summary       = gem.description.dup
-  gem.homepage      = 'https://github.com/kachick/net-ipaddress'
+  gem.summary       = %q{Utils for ipaddress}
+  gem.description   = <<-'DESCRIPTION'
+    ipaddress utilities (Currently supporting only for V4)
+  DESCRIPTION
+  gem.homepage      = repository_url
   gem.license       = 'MIT'
-  gem.name          = 'net-ipaddress'
-  gem.require_paths = ['lib']
-  gem.version       = Net::IPAddress::VERSION.dup
+  gem.name          = lib_name
+  gem.version       = Net::IPAddress::VERSION
 
-  gem.add_development_dependency 'declare', '~> 0.0.6'
-  gem.add_development_dependency 'yard', '>= 0.9.20', '< 2'
-  gem.add_development_dependency 'rake', '>= 10', '< 20'
-  gem.add_development_dependency 'bundler', '>= 2', '< 3'
+  gem.metadata = {
+    'documentation_uri' => 'https://kachick.github.io/net-ipaddress/',
+    'homepage_uri'      => repository_url,
+    'source_code_uri'   => repository_url,
+    'bug_tracker_uri'   => "#{repository_url}/issues"
+  }
 
-  if RUBY_ENGINE == 'rbx'
-    gem.add_dependency 'rubysl', '~> 2.1'
-  end
+  gem.required_ruby_version = Gem::Requirement.new('>= 2.6.0')
 
   # common
 
   gem.authors       = ['Kenichi Kamiya']
   gem.email         = ['kachick1+ruby@gmail.com']
-  gem.files         = `git ls-files`.split($\)
-  gem.executables   = gem.files.grep(%r{^bin/}).map{ |f| File.basename(f) }
-  gem.test_files    = gem.files.grep(%r{^(test|spec|features)/})
+  git_managed_files = `git ls-files`.lines.map(&:chomp)
+  might_be_parsing_by_tool_as_dependabot = git_managed_files.empty?
+  base_files = Dir['README*', '*LICENSE*',  'lib/**/*', 'sig/**/*'].uniq
+  files = might_be_parsing_by_tool_as_dependabot ? base_files : (base_files & git_managed_files)
+
+  unless might_be_parsing_by_tool_as_dependabot
+    if files.grep(%r!\A(?:lib|sig)/!).size < 5
+      raise "obvious mistaken in packaging files, looks shortage: #{files.inspect}"
+    end
+  end
+
+  gem.files         = files
   gem.require_paths = ['lib']
 end
