--- test/declare_helper.rb
@@ -1,5 +1,0 @@
-require 'declare'
-
-$VERBOSE = true
-
-require_relative '../lib/net/ipaddress'
